# src/main/java/org/example/models/Question.java
package org.example.models;

import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class Question {
    private Integer id;
    private String type;
    private String category;
    private String question;
    private List<String> options;
    private Integer number;
}