# src/main/java/org/example/models/QuestionAnswer.java
package org.example.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class QuestionAnswer {
    private Integer questionId;
    private String correctAnswer;
}